setup-all.sh: add asking for sudo at the beggining
This will ask the sudo password once
and keep updating the cache in the background
process once at 60 seconds.

In this way we avoid having to input sudo
later on and probably timing out as this is
a long running script and the user is probably
not looking at it after a while.