ks.cfg: fix interfaces not configured at boot by default
To fix this issue, we used to have a rc.local
script that called dhclient -v. This would initialize
interfaces not up until that point.

However, as this was not managed by systemd, and
likely executed after a lot of networking stuff,
units that depended on network-online (Wants/RequiredBy)
did not work as expected as it didn't have a valid IP
up to that point.

The correct solution is to enable the interfaces to
boot by default by setting ONBOOT=yes in ifcfg
scripts.